backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
psycopg[binary]==3.2.3
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.27.2
